+++ include/soundex.h
@@ -33,6 +33,6 @@
  * Returned char* must be manually freed
  */
 char *soundex(const char *_str);
-float soundex_similarity(const char *str1, const char *str2);
+double soundex_similarity(const char *str1, const char *str2);
 
 #endif /* SOUNDEX_H_ */
